+++ test/libsolidity/ABIJson/inherited.sol
@@ -1,10 +1,10 @@
 contract Base {
-    function baseFunction(uint p) public returns (uint i) { return p; }
+    function baseFunction(uint256 p) public returns (uint256 i) { return p; }
     event baseEvent(bytes32 indexed evtArgBase);
 }
 contract Derived is Base {
     function derivedFunction(bytes32 p) public returns (bytes32 i) { return p; }
-    event derivedEvent(uint indexed evtArgDerived);
+    event derivedEvent(uint256 indexed evtArgDerived);
 }
 // ----
 //     :Base
