+++ test/libsolidity/semanticTests/libraries/internal_library_function_bound_to_external_function.sol
@@ -1,18 +1,18 @@
 library L {
     // NOTE: External function takes up two stack slots
-    function double(function(uint) external pure returns (uint) f, uint x) internal pure returns (uint) {
+    function double(function(uint256) external pure returns (uint256) f, uint256 x) internal pure returns (uint256) {
         return f(x) * 2;
     }
 }
 
 contract C {
-    using L for function(uint) external pure returns (uint);
+    using L for function(uint256) external pure returns (uint256);
 
-    function identity(uint x) external pure returns (uint) {
+    function identity(uint256 x) external pure returns (uint256) {
         return x;
     }
 
-    function test(uint value) public returns (uint) {
+    function test(uint256 value) public returns (uint256) {
         return this.identity.double(value);
     }
 }
