+++ test/libsolidity/semanticTests/viaYul/array_2d_new.sol
@@ -1,8 +1,8 @@
 contract C {
-	function f(uint n) public pure returns (uint) {
-		uint[][] memory a = new uint[][](2);
-		for (uint i = 0; i < 2; ++i)
-			a[i] = new uint[](3);
+	function f(uint256 n) public pure returns (uint256) {
+		uint256[][] memory a = new uint256[][](2);
+		for (uint256 i = 0; i < 2; ++i)
+			a[i] = new uint256[](3);
 		return a[0][0] = n;
 	}
 }
