+++ test/libsolidity/smtCheckerTests/simple/static_array.sol
@@ -2,8 +2,8 @@ pragma experimental SMTChecker;
 contract C
 {
 	// Used to crash because Literal had no type
-	int[3] d;
+	int256[3] d;
 	// Used to crash because Literal had no type
-	int[3*1] x;
+	int256[3*1] x;
 }
 // ----
